refactoring structure for sdks
