+++ cde/sdk/go/go.mod
@@ -1,0 +1,11 @@
+module github.com/cdfoundation/sig-events/cde
+
+go 1.16
+
+require (
+	github.com/cloudevents/sdk-go/v2 v2.3.1
+	github.com/mitchellh/go-homedir v1.1.0
+	github.com/satori/go.uuid v1.2.0
+	github.com/spf13/cobra v1.1.1
+	github.com/spf13/viper v1.7.1
+)
