+++ cde/sdk/go/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	cmd "github.com/cdfoundation/sig-events/cde/cmd"
+)
+func main() {
+	cmd.Execute()
+}
+
