--- cloudevents/go/cde/.goreleaser.yml
@@ -1,73 +1,0 @@
-env:
-  - GO111MODULE=on
-  - CGO_ENABLED=0
-before:
-  hooks:
-    - go mod download
-
-builds:
-  - id: cde
-    # Path to main.go file or main package.
-    # Default is `.`.
-    main: ./cde.go
-
-    # Binary name.
-    # Can be a path (e.g. `bin/app`) to wrap the binary in a directory.
-    # Default is the name of the project directory.
-    binary: cde
-
-    # Custom ldflags templates.
-    # Default is `-s -w -X main.version={{.Version}} -X main.commit={{.ShortCommit}} -X main.date={{.Date}} -X main.builtBy=goreleaser`.
-    ldflags:
-      - -X "{{.Env.ROOTPACKAGE}}/cmd/version.Version={{.Env.VERSION}}" -X "{{.Env.ROOTPACKAGE}}/cmd/version.Revision={{.Env.REV}}" -X "{{.Env.ROOTPACKAGE}}/cmd/version.Branch={{.Env.BRANCH}}" -X "{{.Env.ROOTPACKAGE}}/cmd/version.BuildDate={{.Env.BUILDDATE}}" -X "{{.Env.ROOTPACKAGE}}/cmd/version.GoVersion={{.Env.GOVERSION}}"
-
-    # GOOS list to build for.
-    # For more info refer to: https://golang.org/doc/install/source#environment
-    # Defaults are darwin and linux.
-    goos:
-      - windows
-      - darwin
-      - linux
-
-    # GOARCH to build for.
-    # For more info refer to: https://golang.org/doc/install/source#environment
-    # Defaults are 386 and amd64.
-    goarch:
-      - amd64
-      - arm
-      - arm64
-
-archives:
-  - name_template: "cdf-{{ .Os }}-{{ .Arch }}"
-    format_overrides:
-      - goos: windows
-        format: zip
-
-checksum:
-  # You can change the name of the checksums file.
-  # Default is `cdf_{{ .Version }}_checksums.txt`.
-  name_template: "cde-checksums.txt"
-
-  # Algorithm to be used.
-  # Accepted options are sha256, sha512, sha1, crc32, md5, sha224 and sha384.
-  # Default is sha256.
-  algorithm: sha256
-
-changelog:
-  # set it to true if you wish to skip the changelog generation
-  skip: true
-
-release:
-  # If set to true, will not auto-publish the release.
-  # Default is false.
-  draft: false
-
-  # If set to auto, will mark the release as not ready for production
-  # in case there is an indicator for this in the tag e.g. v1.0.0-rc1
-  # If set to true, will mark the release as not ready for production.
-  # Default is false.
-  prerelease: false
-
-  # You can change the name of the GitHub release.
-  # Default is `{{.Tag}}`
-  name_template: "{{.Env.VERSION}}"
