--- cloudevents/go/cde/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	cmd "github.com/cdfoundation/sig-events/cde/cmd"
-)
-func main() {
-	cmd.Execute()
-}
-
